Not quite as good as I had hoped
This morning I thought we were golden (and we are on the plotting front), but now I've found that the delaunay part is not triangulating the Si well at all for the TestImage.jpg (like 73 found for the whole image).  I suspect a threshold error.  Also hole_dist is not correlated with the Si, maybe with the centers.  Looks like we're setting up for a fun friday :|