Fixed a Deprecation Warning
Not a large change, just fixed a function change that was deprecated after we update python, anaconda and spyder.  With these new updates, Spyder gives a behemoth of a warning/error, but everything goes on pretty much as planned.  Only issue is the toolbar is very ugly now-might have something to do with Tk vs TkAgg?