25% Rewritten
About 25% of the code has been rewritten in an object orientated fashion.

Please note that some lines of commented out code or things marked as 'experimental' have been removed to clean up the code appearence. They will still be present in the older versions of the code if wanted in the future.